Fix PERM issue when restoring an app with redis
This is the root cause of 926224bd5d50f76889f9b7681b0d3d958f100921.

0488aada9f1a0f832a6cf4039db4363e40da2312 caused a regression where
we setup addons in the restore code path. This meant that redis was
instantiated and running, changing the perms of the dir. Then when
the backup extract happens it fails as it has wrong permissions.